test: split preformatted PDF visual coverage
